# services/flights-v2/app/swagger/openapi.json
{"openapi":"3.0.3","info":{"title":"Bee Travels Flights V2 Service","version":"1.0.0","description":"replace this with a real description of your service","license":{"name":"Apache 2.0","url":"https://github.com/bee-travels/bee-travels-python/blob/master/LICENSE"}},"paths":{"/live":{"get":{"description":"Liveness check to make sure service is available","responses":{"200":{"description":"OK"}}}},"/ready":{"get":{"description":"Readiness check to make sure service and all connected external service calls are available","responses":{"200":{"description":"OK"},"503":{"description":"Service Unavailable"}}}},"/api/v1/flights":{"get":{"description":"Example route","responses":{"200":{"description":"OK"},"400":{"description":"Error"}}}},"/api/v1/flights/load":{"get":{"description":"Example route","responses":{"200":{"description":"OK"},"400":{"description":"Error"}}}},"/api/v1/flights/info/{filter}":{"get":{"description":"Get info about flights","parameters":[{"name":"filter","in":"path","description":"info to look up","required":true}],"responses":{"200":{"description":"OK"},"400":{"description":"Error"}}}},"/api/v1/flights/airports":{"get":{"description":"Get all airports","parameters":[{"name":"country","in":"query","description":"Country of the rental company using slug casing (ex. united-states)","required":false,"schema":{"type":"string"}},{"name":"city","in":"query","description":"City of the rental company using slug casing (ex. new-york)","required":false,"schema":{"type":"string"}},{"name":"code","in":"query","description":"3 Letter iata code for the airport","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"},"400":{"description":"Error"}}}},"/api/v1/flights/airports/all":{"get":{"description":"Get all cities and countries we have flights from","responses":{"200":{"description":"OK"},"400":{"description":"Error"}}}},"/api/v1/flights/airports/{id}":{"get":{"description":"Get airport details for a specific airport","parameters":[{"name":"id","in":"path","description":"ID of a airport","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"},"400":{"description":"Error"}}}},"/api/v1/flights/direct/{from}/{to}":{"get":{"description":"Get all direct flight to destination","parameters":[{"name":"from","in":"path","description":"source airport id","required":true},{"name":"to","in":"path","description":"destination airport id","required":true},{"name":"dateFrom","in":"query","description":"Date From","required":true,"schema":{"type":"string"}},{"name":"dateTo","in":"query","description":"Date To","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"},"400":{"description":"Error"}}}},"/api/v1/flights/onestop/{from}/{to}":{"get":{"description":"Get all one stop flight to destination","parameters":[{"name":"from","in":"path","description":"source airport id","required":true},{"name":"to","in":"path","description":"destination airport id","required":true},{"name":"dateFrom","in":"query","description":"Date From","required":true,"schema":{"type":"string"}},{"name":"dateTo","in":"query","description":"Date To","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"},"400":{"description":"Error"}}}},"/api/v1/flights/twostop/{from}/{to}":{"get":{"description":"Get all two stop flight to destination","parameters":[{"name":"from","in":"path","description":"source airport id","required":true},{"name":"to","in":"path","description":"destination airport id0","required":true},{"name":"dateFrom","in":"query","description":"Date From","required":true,"schema":{"type":"string"}},{"name":"dateTo","in":"query","description":"Date To","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK"},"400":{"description":"Error"}}}}},"components":{}}